Transfers ReferableUtil to Referable composable
