+++ aas-web-ui/src/components/Plugins/Submodels/ContactInformations_v1_0.vue
@@ -90,11 +90,11 @@
 <script lang="ts">
     import { defineComponent } from 'vue';
     import { useTheme } from 'vuetify';
+    import { useReferable } from '@/composables/AAS/Referable';
     import { useSMHandling } from '@/composables/SMHandling';
     import RequestHandling from '@/mixins/RequestHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
     import { useAASStore } from '@/store/AASDataStore';
-    import { checkIdShort, descriptionToDisplay, nameToDisplay } from '@/utils/ReferableUtils';
 
     export default defineComponent({
         name: 'ContactInformation',
@@ -106,6 +106,7 @@
             const theme = useTheme();
             const aasStore = useAASStore();
             const { calculateSMEPathes } = useSMHandling();
+            const { checkIdShort, descriptionToDisplay, nameToDisplay } = useReferable();
 
             return {
                 theme, // Theme Object
