+++ aas-web-ui/src/components/Plugins/Submodels/Nameplate_v2_0.vue
@@ -210,11 +210,11 @@
     import { latLng } from 'leaflet';
     import { defineComponent } from 'vue';
     import { useTheme } from 'vuetify';
+    import { useReferable } from '@/composables/AAS/Referable';
     import { useSMHandling } from '@/composables/SMHandling';
     import RequestHandling from '@/mixins/RequestHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
     import { useAASStore } from '@/store/AASDataStore';
-    import { checkIdShort, descriptionToDisplay, nameToDisplay } from '@/utils/ReferableUtils';
 
     export default defineComponent({
         name: 'DigitalNameplate',
@@ -232,6 +232,7 @@
             const theme = useTheme();
             const aasStore = useAASStore();
             const { calculateSMEPathes } = useSMHandling();
+            const { checkIdShort, descriptionToDisplay, nameToDisplay } = useReferable();
 
             return {
                 theme, // Theme Object
