+++ aas-web-ui/src/components/Plugins/Submodels/TechnicalData_v1_2.vue
@@ -254,11 +254,11 @@
 <script lang="ts">
     import { defineComponent } from 'vue';
     import { useTheme } from 'vuetify';
+    import { useReferable } from '@/composables/AAS/Referable';
     import { useSMHandling } from '@/composables/SMHandling';
     import RequestHandling from '@/mixins/RequestHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
     import { useAASStore } from '@/store/AASDataStore';
-    import { checkIdShort, descriptionToDisplay, nameToDisplay } from '@/utils/ReferableUtils';
 
     export default defineComponent({
         name: 'TechnicalData',
@@ -270,6 +270,7 @@
             const theme = useTheme();
             const aasStore = useAASStore();
             const { calculateSMEPathes } = useSMHandling();
+            const { checkIdShort, descriptionToDisplay, nameToDisplay } = useReferable();
 
             return {
                 theme, // Theme Object
