+++ aas-web-ui/src/components/UIComponents/GenericDataTableView.vue
@@ -47,9 +47,9 @@
 <script lang="ts">
     import { defineComponent } from 'vue';
     import { useTheme } from 'vuetify';
+    import { useReferable } from '@/composables/AAS/Referable';
     import { useConceptDescriptionHandling } from '@/composables/ConceptDescriptionHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
-    import { descriptionToDisplay, nameToDisplay } from '@/utils/ReferableUtils';
 
     export default defineComponent({
         name: 'GenericDataTableView',
@@ -72,6 +72,7 @@
         setup() {
             const theme = useTheme();
             const { cdDefinition } = useConceptDescriptionHandling();
+            const { descriptionToDisplay, nameToDisplay } = useReferable();
 
             return {
                 theme, // Theme Object
