+++ aas-web-ui/src/components/UIComponents/GenericDataVisu.vue
@@ -70,10 +70,10 @@
 <script lang="ts">
     import { defineComponent } from 'vue';
     import { useTheme } from 'vuetify';
+    import { useReferable } from '@/composables/AAS/Referable';
     import RequestHandling from '@/mixins/RequestHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
     import { useAASStore } from '@/store/AASDataStore';
-    import { nameToDisplay } from '@/utils/ReferableUtils';
 
     export default defineComponent({
         name: 'GenericDataVisu',
@@ -83,6 +83,7 @@
         setup() {
             const theme = useTheme();
             const aasStore = useAASStore();
+            const { nameToDisplay } = useReferable();
 
             return {
                 theme, // Theme Object
