+++ aas-web-ui/src/components/UIComponents/SubmodelElementGroup.vue
@@ -258,10 +258,10 @@
 // TODO Transfer to composition API
 <script lang="ts">
     import { defineComponent } from 'vue';
+    import { useReferable } from '@/composables/AAS/Referable';
     import { useConceptDescriptionHandling } from '@/composables/ConceptDescriptionHandling';
     import SubmodelElementHandling from '@/mixins/SubmodelElementHandling';
     import { useAASStore } from '@/store/AASDataStore';
-    import { nameToDisplay } from '@/utils/ReferableUtils';
     import DescriptionElement from './DescriptionElement.vue';
 
     export default defineComponent({
@@ -275,6 +275,7 @@
         setup() {
             const aasStore = useAASStore();
             const { unitSuffix } = useConceptDescriptionHandling();
+            const { nameToDisplay } = useReferable();
 
             return {
                 aasStore, // AASStore Object
