+++ aas-web-ui/tests/composables/AAS/Referable.test.ts
@@ -1,5 +1,11 @@
+import { createPinia, setActivePinia } from 'pinia';
 import { describe, expect, it } from 'vitest';
-import { checkIdShort, descriptionToDisplay, nameToDisplay } from '@/utils/ReferableUtils';
+import { useReferable } from '@/composables/AAS/Referable';
+
+const pinia = createPinia();
+setActivePinia(pinia); // Activate Pinia for the test environment
+
+const { checkIdShort, descriptionToDisplay, nameToDisplay } = useReferable();
 
 describe("ReferableUtils.ts; Tests for 'nameToDisplay()'", () => {
     // Define test data for nameToDisplay()
